moved ParamaterPopup to external_gui module
